--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mainRobotPrograms/AutonomousBlueCapBall.java
@@ -1,35 +1,0 @@
-package org.firstinspires.ftc.teamcode.mainRobotPrograms;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-@Autonomous(name = "Autonomous - Blue Cap Ball Edition", group = "Autonomous Group")
-//@Disabled
-
-public class AutonomousBlueCapBall extends _AutonomousBase
-{
-    //Autonomous code for the Blue alliance
-
-    //Called after runOpMode() has finished initializing.
-    protected void driverStationSaysGO() throws InterruptedException
-    {
-        zeroHeading();
-        setMovementPower(-0.5);
-        while (touchSensor.isPressed() == false)
-            updateMotorPowersBasedOnGyroHeading();
-        stopDriving();
-        adjustHeading();
-
-        setPrecisionFactor(5);
-        turnToHeading(-105, TurnMode.BOTH);
-
-        zeroHeading();
-        setMovementPower(0.5);
-        while (bottomColorSensor.blue() <= 2)
-            updateMotorPowersBasedOnGyroHeading();
-        stopDriving();
-
-        harvester.setPower(-0.5);
-        sleep(2000);
-        harvester.setPower(0);
-    }
-}
